intel_iommu: fix some gpasid related issues
During tests, we found the NIC pf/vf (SRIOV) passthrough
cannot work correctly. The NIC cannot get IP in VM. One
reason is that the qemu execute UNBIND action after BIND
for this device. So, we should not execute UNBIND if the
device is a passthrough device and its gpasid is already
bound.

Signed-off-by: Yi Sun <yi.y.sun@linux.intel.com>